feat: move tce status -> node status
